Complete refactoring of the plugin to a new, more modular and maintainable structure.
This commit completes the major refactoring of the plugin architecture. The codebase is now aligned with a modern, namespaced, and organized structure following the defined schema.

Key changes:
- A new directory structure has been created under `includes/` to separate concerns (Admin, Api, Data, Services, Public, etc.).
- A PSR-4 autoloader has been added to handle class loading.
- The main plugin file `kerbcycle-qr-code-manager.php` has been stripped down to a bootstrap loader.
- A new `includes/Plugin.php` class now orchestrates the plugin's components.
- All logic from the original monolithic files has been broken down and moved into dedicated classes within the new structure.
- All original class files have been refactored and removed.
- An additional REST API endpoint from an untracked file (`class-kerbcycle-qr-api.php`) has been integrated into the new structure.
- The notification functionality, which was broken during the refactoring, has been restored by implementing the `EmailService` and `SmsService` and updating the AJAX handlers to use them.
- The plugin's code is now organized, more maintainable, and follows modern PHP standards.
