+++ assets/css/admin.css
@@ -1,0 +1,1 @@
+/* Styles for the admin pages */
