+++ assets/css/public.css
@@ -1,0 +1,1 @@
+/* Styles for the public pages */
