+++ assets/js/admin.js
@@ -1,0 +1,1 @@
+// JavaScript for the admin pages
