+++ includes/Data/Repositories/QrCodeRepository.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Kerbcycle\QrCode\Data\Repositories;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+/**
+ * The qr code repository.
+ *
+ * @since      1.0.0
+ * @package    Kerbcycle\QrCode
+ * @subpackage Kerbcycle\QrCode\Data\Repositories
+ */
+class QrCodeRepository
+{
+    private $table;
+
+    public function __construct()
+    {
+        global $wpdb;
+        $this->table = $wpdb->prefix . 'kerbcycle_qr_codes';
+    }
+
+    // Methods for CRUD operations on qr_codes table will be added here.
+}
