+++ includes/Data/Tables.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Kerbcycle\QrCode\Data;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+/**
+ * The tables.
+ *
+ * @since      1.0.0
+ * @package    Kerbcycle\QrCode
+ * @subpackage Kerbcycle\QrCode\Data
+ */
+class Tables
+{
+    // This class will be responsible for creating and managing the custom database tables.
+}
