+++ includes/Helpers/Nonces.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Kerbcycle\QrCode\Helpers;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+/**
+ * The nonces helper.
+ *
+ * @since      1.0.0
+ * @package    Kerbcycle\QrCode
+ * @subpackage Kerbcycle\QrCode\Helpers
+ */
+class Nonces
+{
+    // Methods for creating and verifying nonces will be added here.
+}
