+++ includes/Helpers/Utils.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Kerbcycle\QrCode\Helpers;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+/**
+ * The utils helper.
+ *
+ * @since      1.0.0
+ * @package    Kerbcycle\QrCode
+ * @subpackage Kerbcycle\QrCode\Helpers
+ */
+class Utils
+{
+    // Utility functions will be added here.
+}
