+++ includes/Install/Uninstaller.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Kerbcycle\QrCode\Install;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+/**
+ * Fired during plugin deactivation.
+ *
+ * This class defines all code necessary to run during the plugin's deactivation.
+ *
+ * @since      1.0.0
+ * @package    Kerbcycle\QrCode
+ * @subpackage Kerbcycle\QrCode\Install
+ */
+class Uninstaller
+{
+    /**
+     * Short Description. (use period)
+     *
+     * Long Description.
+     *
+     * @since    1.0.0
+     */
+    public static function deactivate()
+    {
+        // Optional: Add cleanup logic here.
+    }
+}
