+++ includes/Services/QrService.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Kerbcycle\QrCode\Services;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+use Kerbcycle\QrCode\Data\Repositories\QrCodeRepository;
+
+/**
+ * The qr service.
+ *
+ * @since      1.0.0
+ * @package    Kerbcycle\QrCode
+ * @subpackage Kerbcycle\QrCode\Services
+ */
+class QrService
+{
+    private $repository;
+
+    public function __construct()
+    {
+        $this->repository = new QrCodeRepository();
+    }
+
+    // Methods for assigning, releasing, updating, finding, and bulk actions will be added here.
+}
